At e2e test for hot-reload
